pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "roninhand-hri-bridge"
version = "0.1.0"
description = "Input-to-gesture bridge that maps EMG/EEG/IMU/VR intents to RoninHand gestures via HTTP."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name="NatIgnite 2025 Team"}]
dependencies = [
    "requests>=2.32",
    "numpy>=1.26",
    "scipy>=1.12",
    "pyyaml>=6.0.1",
    "pydantic>=2.6",
    "rich>=13.7",
]

[project.optional-dependencies]
brainflow = ["brainflow>=5.13.0"]
dev = ["pytest>=8.0", "ruff>=0.4.5", "black>=24.4.2", "mypy>=1.10"]

[project.scripts]
rhbridge = "rhbridge.cli:main"
